+++ Dockerfile
@@ -35,9 +35,7 @@ RUN GOOS=${TARGETOS} GOARCH=${TARGETARCH} go build \
       -ldflags "-w -X ${VERSION_PACKAGE}.version=${VERSION} -X ${VERSION_PACKAGE}.buildDate=$(date -u +'%Y-%m-%dT%H:%M:%SZ') -X ${VERSION_PACKAGE}.gitCommit=${GIT_COMMIT} -X ${VERSION_PACKAGE}.gitTreeState=${GIT_TREE_STATE}" \
       -o bin/kargo-render \
       ./cmd \
-    && bin/kargo-render version \
-    && cd bin \
-    && ln -s kargo-render kargo-render-action
+    && bin/kargo-render version
 
 FROM alpine:3.19.1 as final
