--- cmd/action/logger.go
@@ -1,5 +1,0 @@
-package action
-
-import libLog "github.com/akuity/kargo-render/internal/log"
-
-var logger = libLog.LoggerOrDie()
