+++ cmd/action_cmd_test.go
@@ -1,4 +1,4 @@
-package action
+package main
 
 import (
 	"fmt"
