+++ cmd/output.go
@@ -1,4 +1,4 @@
-package cli
+package main
 
 import (
 	"encoding/json"
